Move AuthenticationManager.h to NetworkProcess directory https://bugs.webkit.org/show_bug.cgi?id=283997 rdar://140872285
Reviewed by Tim Horton.

There was a time when we did authentication in the web content process,
but that was removed years ago.  This just moves the files, removes unnecessary
header includes, and fixes the unified build.

* Source/WebKit/CMakeLists.txt:
* Source/WebKit/DerivedSources-input.xcfilelist:
* Source/WebKit/DerivedSources.make:
* Source/WebKit/NetworkProcess/Authentication/AuthenticationManager.cpp: Renamed from Source/WebKit/Shared/Authentication/AuthenticationManager.cpp.
(WebKit::AuthenticationManager::Challenge::Challenge):
(WebKit::canCoalesceChallenge):
(WebKit::AuthenticationManager::supplementName):
(WebKit::AuthenticationManager::AuthenticationManager):
(WebKit::AuthenticationManager::ref const):
(WebKit::AuthenticationManager::deref const):
(WebKit::AuthenticationManager::protectedProcess const):
(WebKit::AuthenticationManager::addChallengeToChallengeMap):
(WebKit::AuthenticationManager::shouldCoalesceChallenge const):
(WebKit::AuthenticationManager::coalesceChallengesMatching const):
(WebKit::AuthenticationManager::didReceiveAuthenticationChallenge):
(WebKit::AuthenticationManager::completeAuthenticationChallenge):
(WebKit::AuthenticationManager::negotiatedLegacyTLS const):
* Source/WebKit/NetworkProcess/Authentication/AuthenticationManager.h: Renamed from Source/WebKit/Shared/Authentication/AuthenticationManager.h.
* Source/WebKit/NetworkProcess/Authentication/AuthenticationManager.messages.in: Renamed from Source/WebKit/Shared/Authentication/AuthenticationManager.messages.in.
* Source/WebKit/NetworkProcess/Authentication/cocoa/AuthenticationManagerCocoa.mm: Renamed from Source/WebKit/Shared/Authentication/cocoa/AuthenticationManagerCocoa.mm.
(WebKit::AuthenticationManager::initializeConnection):
* Source/WebKit/NetworkProcess/cocoa/NetworkTaskCocoa.mm:
* Source/WebKit/NetworkProcess/webtransport/cocoa/NetworkTransportReceiveStreamCocoa.mm:
* Source/WebKit/Shared/Cocoa/RevealItem.h:
* Source/WebKit/Sources.txt:
* Source/WebKit/SourcesCocoa.txt:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/287310@main
