+++ Source/WebKit/Sources.txt
@@ -123,6 +123,8 @@ NetworkProcess/NetworkSocketChannel.cpp
 NetworkProcess/PingLoad.cpp
 NetworkProcess/PreconnectTask.cpp
 
+NetworkProcess/Authentication/AuthenticationManager.cpp
+
 NetworkProcess/Classifier/ITPThirdPartyData.cpp
 NetworkProcess/Classifier/ITPThirdPartyDataForSpecificFirstParty.cpp
 NetworkProcess/Classifier/ResourceLoadStatisticsStore.cpp
@@ -299,8 +301,6 @@ Shared/API/c/WKURL.cpp
 Shared/API/c/WKURLRequest.cpp
 Shared/API/c/WKURLResponse.cpp
 Shared/API/c/WKUserContentURLPattern.cpp
-
-Shared/Authentication/AuthenticationManager.cpp
 
 Shared/Databases/IndexedDB/IDBUtilities.cpp
