+++ Source/WebKit/SourcesCocoa.txt
@@ -21,6 +21,8 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
 // THE POSSIBILITY OF SUCH DAMAGE.
 
+NetworkProcess/Authentication/cocoa/AuthenticationManagerCocoa.mm
+
 NetworkProcess/cache/NetworkCacheDataCocoa.mm
 NetworkProcess/cache/NetworkCacheIOChannelCocoa.mm
 
@@ -156,7 +158,6 @@ Shared/WebSQLiteDatabaseTracker.cpp
 Shared/cf/CookieStorageUtilsCF.mm
 
 Shared/Authentication/cocoa/AuthenticationChallengeDispositionCocoa.mm
-Shared/Authentication/cocoa/AuthenticationManagerCocoa.mm
 Shared/Authentication/cocoa/ClientCertificateAuthenticationXPCConstants.cpp
 
 Shared/EntryPointUtilities/Cocoa/Daemon/PCMDaemonConnectionSet.mm
